refactor(openai-models): re-export Usage / *TokensDetails from canonical types
src/core/models/openai/responses.rs declared its own Usage,
PromptTokensDetails, and CompletionTokensDetails with field-for-field
identical layouts to the canonical core::types::responses::*. The
duplicates were a downstream of the parallel-type-tree schism (#519,
audit A-2/A-5).

Replace the duplicate struct definitions with a `pub use` re-export
of the canonical types. All existing callers — chat.rs constructors,
streaming/handler, streaming/types, tests — continue to compile
unchanged because the type identity now resolves to the same struct.

Three structural notes:

1. The 966-line responses.rs was over the U-16 800-line ceiling, so
   the inline test module (681 lines) was first split out to
   responses_tests.rs via `#[cfg(test)] #[path = "..."] mod tests;`,
   bringing the main file to 248 lines. Same pattern PR #531 used
   for llm_cache.rs.

2. core::types::responses::Usage skips serializing None on the
   Option fields more aggressively than the old openai-models copy
   did. Net wire effect: a `null` cached_tokens / null reasoning_tokens
   in the response body is now omitted instead of written. Stricter
   wire shape, OpenAI clients accept either form.

3. The unused `use crate::core::types::thinking::ThinkingUsage;`
   import is removed from the main file (it was only needed for the
   removed Usage struct's field type) and added to responses_tests.rs
   where the test module still constructs ThinkingUsage instances.

Architecture roadmap context (#519): second slice of Phase 2 type
collapse, after PR #534 deleted the entire dead transform/ tree.
4 of the original 6 Usage definitions remain (sdk, pricing, langfuse,
core::types::responses canonical) — sdk/langfuse have different
schemas/wire-formats and are not candidates; pricing::Usage is a
mechanical replacement across 15+ providers, scheduled as a separate
PR-A3.

Net: 10333 lib tests pass.
